+++ tests/Medidata.MAuth.Tests/Infrastructure/MAuthServerHandler.cs
@@ -4,26 +4,45 @@ using System.Net.Http;
 using System.Threading;
 using System.Threading.Tasks;
 using Medidata.MAuth.Core;
-using Microsoft.Extensions.Logging.Abstractions;
+using Medidata.MAuth.Tests.ProtocolTestSuite;
 using Newtonsoft.Json;
 
 namespace Medidata.MAuth.Tests.Infrastructure
 {
     internal class MAuthServerHandler : HttpMessageHandler
     {
-        private static readonly Guid clientUuid = new Guid("192cce84-8466-490e-b03e-074f82da3ee2");
-        private int currentNumberOfAttempts = 0;
+        private MAuthServerHandler() { }
+
+        private static readonly Guid _clientUuid = new Guid("192cce84-8466-490e-b03e-074f82da3ee2");
+        private int _currentNumberOfAttempts = 0;
 
         public int SucceedAfterThisManyAttempts { get; set; } = 1;
+
+        private static string _signingPublicKey;
+        private static Guid _signingAppUuid;
+
+        private async Task<MAuthServerHandler> InitializeAsync()
+        {
+            var protocolSuite = new ProtocolTestSuiteHelper();
+            _signingPublicKey = await protocolSuite.GetPublicKey();
+            _signingAppUuid = await protocolSuite.ReadSignInAppUuid();
+            return this;
+        }
+
+        public static Task<MAuthServerHandler> CreateAsync()
+        {
+            var ret = new MAuthServerHandler();
+            return ret.InitializeAsync();
+        }
 
         protected override async Task<HttpResponseMessage> SendAsync(
             HttpRequestMessage request, CancellationToken cancellationToken)
         {
-            currentNumberOfAttempts += 1;
+            _currentNumberOfAttempts += 1;
             var version = request.GetAuthHeaderValue().GetVersionFromAuthenticationHeader();
             var mAuthCore = MAuthCoreFactory.Instantiate(version);
 
-            if (currentNumberOfAttempts < SucceedAfterThisManyAttempts)
+            if (_currentNumberOfAttempts < SucceedAfterThisManyAttempts)
                 return new HttpResponseMessage(HttpStatusCode.ServiceUnavailable);
             var authInfo = MAuthAuthenticator.GetAuthenticationInfo(request, mAuthCore);
 
@@ -34,9 +53,17 @@ namespace Medidata.MAuth.Tests.Infrastructure
                 return new HttpResponseMessage(HttpStatusCode.Unauthorized) { RequestMessage = request };
 
             if (!request.RequestUri.AbsolutePath.Equals(
-                $"{Constants.MAuthTokenRequestPath}{clientUuid.ToHyphenString()}.json",
+                $"{Constants.MAuthTokenRequestPath}{_clientUuid.ToHyphenString()}.json",
+                StringComparison.OrdinalIgnoreCase) &&
+                !request.RequestUri.AbsolutePath.Equals(
+                $"{Constants.MAuthTokenRequestPath}{_signingAppUuid.ToHyphenString()}.json",
                 StringComparison.OrdinalIgnoreCase))
+            {
                 return new HttpResponseMessage(HttpStatusCode.NotFound) { RequestMessage = request };
+            }
+
+            bool isProtocolSuiteTest = request.RequestUri.AbsolutePath.Contains(
+                _signingAppUuid.ToHyphenString());
 
             return new HttpResponseMessage(HttpStatusCode.OK)
             {
@@ -46,10 +73,10 @@ namespace Medidata.MAuth.Tests.Infrastructure
                     {
                         security_token = new ApplicationInfo()
                         {
-                            Uuid = clientUuid,
+                            Uuid = isProtocolSuiteTest ? _signingAppUuid : _clientUuid,
                             Name = "Medidata.MAuth.Tests",
                             CreationDate = new DateTimeOffset(2016, 8, 1, 0, 0, 0, TimeSpan.Zero),
-                            PublicKey = TestExtensions.ClientPublicKey
+                            PublicKey = isProtocolSuiteTest ? _signingPublicKey : TestExtensions.ClientPublicKey
                         }
                     })
                 )